add navigation to age page
