+++ feature/gender/src/main/java/com/dante/calorietracker/feature/gender/navigation/GenderNavigation.kt
@@ -1,0 +1,16 @@
+package com.dante.calorietracker.feature.gender.navigation
+
+import androidx.navigation.NavController
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.NavOptions
+import androidx.navigation.compose.composable
+
+const val genderRoute = "gender_route"
+
+fun NavController.navigateToGender(navOptions: NavOptions? = null) =
+    this.navigate(genderRoute, navOptions)
+
+fun NavGraphBuilder.genderScreen() {
+    composable(route = genderRoute) {
+    }
+}
