--- feature/weight/src/main/java/com/dante/calorietracker/feat/weight/navigation/WeightNavigation.kt
@@ -1,9 +1,0 @@
-package com.dante.calorietracker.feat.weight.navigation
-
-import androidx.navigation.NavController
-import androidx.navigation.NavOptions
-
-const val weightRoute = "weight_route"
-
-fun NavController.navigateToWeight(navOptions: NavOptions? = null) =
-    this.navigate(weightRoute, navOptions)
